# Application/Swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CashFlow.API",
    "version": "v1"
  },
  "paths": {
    "/api/v{version}/CashMoviment/{id}": {
      "get": {
        "tags": [
          "CashMoviment"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "CashMoviment"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/CashMoviment": {
      "post": {
        "tags": [
          "CashMoviment"
        ],
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashMovimentModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CashMovimentModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CashMovimentModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "patch": {
        "tags": [
          "CashMoviment"
        ],
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashMovimentModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CashMovimentModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CashMovimentModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/CashMoviment/balance": {
      "get": {
        "tags": [
          "CashMoviment"
        ],
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "end",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CashMovimentModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createAt": {
            "type": "string",
            "format": "date-time"
          },
          "historic": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "number",
            "format": "double"
          },
          "nature": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CashMovimentModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createAt": {
            "type": "string",
            "format": "date-time"
          },
          "historic": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "number",
            "format": "double"
          },
          "nature": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      }
    }
  }
}